rewrote database operations
Shouldn't have to upgrade the DB now. Upgrading the DB was needed to add more emails in the last method, but it causes issues as the DB will go offline during that process.
There is a bug with getting new image. This is due to image Ids function. Will come back to this.